+++ CHANGELOG.md
@@ -5,6 +5,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [UNRELEASED]
+### Changed
+- New ASCII Logo
+
+### Added
+- Version Command with Logo Output
+
 
 ## [0.0.9] - 2023-02-06
 ### Added